requirements.txt
Django>=4.0,<5.0
gunicorn
whitenoise
dj-database-url
xhtml2pdf
django-allauth